pyproject.toml
[project]
name = "noize"
version = "0.1.0"
description = "A cli tool that uses adversarial perturbation to protect audio clips from being used in unauthorised voice cloning"
readme = "README.md"
requires-python = ">=3.10 <3.12"
dependencies = [
    "ctranslate2 (<=4.4.0)",
    "torch (>=2.2.0,<3.0.0)",
    "typer (>=0.15.1,<0.16.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "coqui-tts (>=0.26.0,<0.27.0)",
    "flask (>=3.1.0,<4.0.0)",
]

[project.scripts]
noize = 'noize.cli.main:app'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
